changes to the mafft and raxml components
