+++ flowcraft/generator/components/metagenomics.py
@@ -346,75 +346,3 @@ class SplitAssembly(Process):
             "split_assembly"
         ]
 
-
-class Mafft(Process):
-    """mafft to align sequences
-
-            This process is set with:
-
-                - ``input_type``: fasta
-                - ``output_type``: align
-                - ``ptype``: sequence alignment
-
-            """
-
-    def __init__(self, **kwargs):
-
-        super().__init__(**kwargs)
-
-        self.input_type = "fasta"
-        self.output_type = "align"
-
-        self.params = {
-        }
-
-        self.directives = {
-            "mafft": {
-                "container": "flowcraft/mafft",
-                "version": "7.402-1",
-                "cpus": 4,
-                "memory": "{ 4.GB * task.attempt }"
-            }
-        }
-
-        self.status_channels = [
-            "mafft"
-        ]
-
-class Raxml(Process):
-    """mafft to align sequences
-
-            This process is set with:
-
-                - ``input_type``: align
-                - ``output_type``: .tree
-                - ``ptype``: tree
-
-            """
-
-    def __init__(self, **kwargs):
-
-        super().__init__(**kwargs)
-
-        self.input_type = "align"
-        self.output_type = ".tree"
-
-        self.params = {
-            "substitutionModel": {
-                "default": "'GTRGAMMA'",
-                "description": "Substitution model. Option: GTRCAT, GTRCATI, ASC_GTRCAT, GTRGAMMA, ASC_GTRGAMMA etc "
-            }
-        }
-
-        self.directives = {
-            "raxml": {
-                "container": "flowcraft/raxml",
-                "version": "8.2.11-1",
-                "cpus": 4,
-                "memory": "{ 4.GB * task.attempt }"
-            }
-        }
-
-        self.status_channels = [
-            "raxml"
-        ]
