+++ flowcraft/generator/templates/mafft.nf
@@ -2,13 +2,15 @@ process mafft_{{ pid }} {
 
     {% include "post.txt" ignore missing %}
 
+    tag { 'mafft' }
+
     publishDir "results/mafft/"
 
     input:
     file(assembly) from {{ input_channel }}.map{ it[1] }.collect()
 
     output:
-    file ("out.align") into {{ output_channel }}
+    file ("*.align") into {{ output_channel }}
     {% with task_name="mafft", sample_id="val('single')" %}
     {%- include "compiler_channels.txt" ignore missing -%}
     {% endwith %}
@@ -18,7 +20,7 @@ process mafft_{{ pid }} {
 
     cat ${assembly} > all_assemblies.fasta
 
-    mafft --thread $task.cpus --auto all_assemblies.fasta > out.align
+    mafft --thread $task.cpus --auto all_assemblies.fasta > ${workflow.scriptName}.align
     """
 
 }
